refactor(admin): AdminVisio sous 300 lignes — éclatement complet (H3)
Éclate AdminVisio.vue (820 l) selon le pattern G1 (orchestrateur ≤300 +
composable + sous-composants présentationnels), à parité de rendu et de
comportement.

- useAdminVisio.js : toute la logique <script> (état, filtres, stats,
  computed, load/process/refresh, helpers statut/format, joinVisio/
  viewVisioDetails, onMounted). Import nommé { klassciService } conservé
  à l'identique (parité avec l'original ; AdminUsers utilise l'import par
  défaut — non aligné volontairement pour ne pas changer le comportement).
- VisioStatsCards.vue : les 4 cartes de stats (prop stats).
- VisioFilters.vue : période/statut/recherche en v-model (defineModel) ;
  émet reload (changement de période) et filter (statut/recherche),
  reproduisant @change=loadVisioconferences et @change|@input=
  filterVisioconferences.
- VisioGrid.vue : grille de cartes ; reçoit la liste filtrée + helpers
  getStatusLabel/formatDate/formatTime en props ; émet join/view.
- AdminVisio.vue : header, états chargement/erreur/vide, pagination, CSS
  résiduel (container/header/states/pagination + responsive container/title).

CSS déplacé VERBATIM dans chaque sous-composant avec son @media : le bloc
responsive d'origine ciblait '.stats-grid, .visio-grid' et '.filters-grid' ;
en scoped, chaque règle suit son markup (stats-grid -> VisioStatsCards,
visio-grid -> VisioGrid, filters-grid -> VisioFilters). Aucune CSS supprimée.

Dette pré-existante conservée à l'identique (non corrigée) :
- joinVisio/viewVisioDetails contiennent des TODO et des console.log.
- filterVisioconferences est un no-op (le filtrage est porté par le computed) ;
  conservé et toujours câblé via l'émission 'filter' pour parité stricte.
- Pas de vraie pagination (juste un compteur d'affichage) — inchangé.

Tests : useAdminVisio.test.js (composable, service+cache mockés) et
VisioGrid.test.js (rendu + émissions) = 11/11 verts. Build vite exit 0.
